Move blocks actions under preferences namespace
