+++ app/views/preferences/blocks/index.html.haml
@@ -1,5 +1,5 @@
 %p.info_message Users that you have blocked will not be able to read your plan, and you will not see each other listed in quicklove or search results.
-= link_to "See the FAQ for more information", about_blocks_path
+= link_to "See the FAQ for more information", about_preferences_blocks_path
 %p
   %b Please be aware that your activity on Notes is still visible to everyone.
   If you feel this presents a serious problem to you, please
